requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
requests==2.31.0
beautifulsoup4==4.12.2
googlemaps==4.10.0
instaloader==4.10.1
facebook-scraper==0.2.59
playwright==1.40.0
python-dotenv==1.0.0
lxml==4.9.3
regex==2023.10.3
Werkzeug==2.3.7
scrapy==2.13.2

# Nouvelles dépendances pour l'analyse IA
openai==1.3.7

# Dépendances de développement et tests
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0 
google-cloud-bigquery 
selenium 